pyproject.toml
[tool.poetry]
name = "kivyopencvfps"
version = "0.1.0"
description = ""
authors = ["AccelQuasarDragon <138998466+AccelQuasarDragon@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
Kivy = "^2.2.1"
opencv-python = "^4.8.0.74"
numpy = "^1.25.1"
pillow = "^10.0.0"
pygame = "^2.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
